hal: add support for config sta v1 command
Mako firmware interface (1.3.1.0) needs a bit different
config sta command. To handle that, create struct
wcn36xx_hal_config_sta_req_msg_v1 which is to be used with mako firmware.
wcn36xx will dynamically detect from firmware API version if older (_v1) or
newer version of the command structure needs to be used.

I implemented it so that we always populate the new structure first and then if
we detect that older structure is needed we just copy the values from the
original structure to the v1 structure. That way structure needs to be
populated only once.

Signed-off-by: Kalle Valo <kvalo@qca.qualcomm.com>