Use locally adminstered mac if file is not available
If there is no mac address file available for all but the first
interface use locally administered mac addresses. This is ok since
they will be used for p2p action where locally administered addresses
are ok. The locally administered addresses will be based on the
address for interface one.

If the mac address file is missing for the first interface we still
randomize one to use.

Signed-off-by: Olof Johansson <dev@skyshaper.net>