Remove firmware allocation from wcn struct
Remove the nv pointer from the wcn struct and keep it inside the
smd_load_nv function as local. This way we don't keep it allocated
throught the life of the module and won't waste memory.

This commit also includes minor cleanup of unused variables in the
smd_load_nv function.

Signed-off-by: Olof Johansson <dev@skyshaper.net>