Fix TX stall in case ring is completely full
In the situation when tx ring is completely full head and tail
are pointing to the same element. Next dxe tx complete interrupt
will not take any element from the ring because it will think
that ring is empty. As a workaround for each dxe tx complete
interrupt take at least one item from the ring.

Signed-off-by: Eugene Krasnikov <k.eugene.e@gmail.com>